messages_relay

Module message_lane_loop

Message delivery loop. Designed to work with messages pallet.

Single relay instance delivers messages of single lane in single direction. To serve two-way lane, you would need two instances of relay. To serve N two-way lanes, you would need N*2 instances of relay.

Please keep in mind that the best header in this file is actually best finalized header. I.e. when talking about headers in lane context, we only care about finalized headers.

Structs

ClientState

State of the client.

ClientsState

Both clients state.

MessageDeliveryParams

Message delivery race parameters.

MessageDetails

Message details.

MessageProofParameters

Message delivery race proof parameters.

NoncesSubmitArtifacts

Artifacts of submitting nonces proof.

Params

Message lane loop configuration params.

Traits

BatchTransaction

Batch transaction that already submit some headers and needs to be extended with messages/delivery proof before sending.

SourceClient

Source client trait.

TargetClient

Target client trait.

Functions

metrics_prefix

Return prefix that will be used by default to expose Prometheus metrics of the finality proofs sync loop.

run

Run message lane service loop.

Type Aliases

MessageDetailsMap

Messages details map.

SourceClientState

State of source client in one-way message lane.

TargetClientState

State of target client in one-way message lane.