Type Alias messages_relay::message_lane_loop::TargetClientState

pub type TargetClientState<P> = ClientState<TargetHeaderIdOf<P>, SourceHeaderIdOf<P>>;

State of target client in one-way message lane.

Aliased Type

struct TargetClientState<P> {
    pub best_self: HeaderId<<P as MessageLane>::TargetHeaderHash, <P as MessageLane>::TargetHeaderNumber>,
    pub best_finalized_self: HeaderId<<P as MessageLane>::TargetHeaderHash, <P as MessageLane>::TargetHeaderNumber>,
    pub best_finalized_peer_at_best_self: Option<HeaderId<<P as MessageLane>::SourceHeaderHash, <P as MessageLane>::SourceHeaderNumber>>,
    pub actual_best_finalized_peer_at_best_self: Option<HeaderId<<P as MessageLane>::SourceHeaderHash, <P as MessageLane>::SourceHeaderNumber>>,
}

Fields
best_self: HeaderId<<P as MessageLane>::TargetHeaderHash, <P as MessageLane>::TargetHeaderNumber>

The best header id of this chain.

best_finalized_self: HeaderId<<P as MessageLane>::TargetHeaderHash, <P as MessageLane>::TargetHeaderNumber>

Best finalized header id of this chain.

best_finalized_peer_at_best_self: Option<HeaderId<<P as MessageLane>::SourceHeaderHash, <P as MessageLane>::SourceHeaderNumber>>

Best finalized header id of the peer chain read at the best block of this chain (at best_finalized_self).

It may be None e,g. if peer is a parachain and we haven’t yet relayed any parachain heads.

actual_best_finalized_peer_at_best_self: Option<HeaderId<<P as MessageLane>::SourceHeaderHash, <P as MessageLane>::SourceHeaderNumber>>

Header id of the peer chain with the number, matching the best_finalized_peer_at_best_self.